Add Minecraft authentication handler and associated DTOs
This update introduces a Minecraft authentication handler, which includes associated data transfer objects for the Metadata and MetadataResponse. Encryption and decryption services have also been added for serving private and public keys. Dockerfile was updated to install openssl for use in the key services. Versatility is further improved with the addition of 'ProjectVersion' and 'SkinDomains' fields in the server settings.