Refactor and update code for integrations and improvements
Multiple changes were made in this commit, including refactoring of the AuthServiceFactory and the addition of new services for SentryErrorSaveHandler and AnyAuthService. The ResponseMessage class was relocated under the 'Dto/Messages' directory. Updates were also made to validation criteria and to error handling. Furthermore, the TextureServiceDto class was renamed to UrlServiceDto.