Refactor AuthType to use enum and enrich API responses
The commit updates `AuthType` in `integrationUpdateDto` from integer type to `AuthType` enum. This simplifies and enhances readability of API codes. Also, the responses from Endpoints are now more informative, as they return a `SettingsReadDto` object. The changes also include necessary updates to the validators and handlers to accommodate this datatype change, thus enhancing the robustness of our validation system.