Update Docker base image and introduce path normalization
This change updates the Docker base image from aspnet to sdk. Also, it introduces a new function to normalize paths in the GitHubService. In addition, a method for creating OK responses in the EndpointHelper has been added. Furthermore, a new method to compile source code in different versions and a method to copy directories have been added in the GitHubLauncherHub. These changes are intended to improve the robustness and functionality of our application.