Reorder test commands in docker packaging workflow
The git diff shows that we've rearranged the execution of test commands in the docker packaging workflow. Now, the Gml Core and Gml Web Api tests run before the installation of the cosign tool, not after as previously. This order could potentially prevent unnecessary cosign installations if the tests fail.