Update FileWhiteListValidator to handle list of FileWhiteListDto
The FileWhiteListValidator class has been updated to handle a list of FileWhiteListDto objects rather than a single object. Correspondingly, changes have been made in the AddFileWhiteList and RemoveFileWhiteList methods as well as in the ValidatorsExtensions class. The duplication of file hashes and profile names in the list of FileWhiteListDtos is handled to prevent redundant operations.