Add connection management for launcher users
The commit involves improvements in connection management of launcher users. A connection attribute has been added in UserLauncherInfo for maintaining the connection state. Code has been added for regular checks for connection status of each launcher user, along with disposing of timer for inactive users. This aids in better management of active users and efficient use of resources.