Update ConfigureWebHost in GmlApiApplicationFactory
The update comprises of a new in-memory configuration dictionary in ConfigureWebHost method of GmlApiApplicationFactory class in the test project. This update refines the services.AddAuthentication code block and introduces a ConfigureAppConfiguration block to add specific configurations. The configuration values are set in an in-memory collection, improving the way settings are managed during testing.