Implement game loader version check in ProfileHandler
The commented code for version checking in ProfileHandler.cs has been uncommented and implemented. It verifies the game loader type, and retrieves the list of allowed versions from the game manager. It then returns these versions in the response or throws a BadRequest result if the game loader cannot be determined.