Rename 'TextureUrl' and Update Gml.Core (#54)
* Rename 'TextureUrl' to 'TextureSkinUrl' in PlayerReadDto

The name of the property 'TextureUrl' has been updated to 'TextureSkinUrl' in PlayerReadDto class to enhance name clarity and better represent its purpose. This change helps improve code readability and maintainability.

* Update Gml.Core in Web API

The Gml.Core file in the Web API directory has been updated. The change in the file enhances the efficiency of the system.