Add NeoForge version installer for the NeoForge project
This notable update introduces the NeoForge version installer for the NeoForge project. A mechanism to extract and install libraries has been added, including the necessary processors and necessary file class constructors. Furthermore, the commit contains the implementation of install options, an important part of the version installer. In addition, changes allow for file downloading and extraction, making the installer ready for use.