Feature/rate limit (#67)
* Add rate limiting to API

Introduced a RateLimitExtension for configuring global rate limiting on the API. Updated ApplicationExtensions to use the rate limiter and simplified GmlManager setup through a new configuration method.

* Add GmlConfigurationExtension for Gml manager configuration

Introduce a new static class `GmlConfigurationExtension` that provides an extension method for configuring `IGmlManager` in `IServiceCollection`. It initializes `GmlManager` with project settings and restores launcher settings.