Update to v0.1.0-rc2-hotfix1 (#80)
* Remove Address and Port properties from ServerReadDto

This commit refines the ServerReadDto model by removing the Address and Port properties. This change simplifies the data transfer object and aligns better with current application requirements.

* Fix endpoint URL in MinecraftHandler

Update the URL for the cape texture from 'skins' to 'capes' to resolve incorrect resource reference. This ensures the proper endpoint is accessed for retrieving the cape textures for users.